Fix for hacky way of deriving acceleration due to gravity.
The sample-data-utility now computes the acceleration due to gravity based on the magnitude of the first accelerometer data sample. The algorithm then multilpies this by the normalized accelerometer data sample for each sample from then on, and subtracts the result of this multiplication from the total acceleration. What is left is the acceleration from all forces minus gravity. This approach assumes the first sample from the IMU represents that the IMU is "motionless". Later, this value can be calibrated and stored, because the acceleration due to gravity for the same IMU should be roughly the same every time.